chore(shared,clerk-js,elements): Move webauthn utilities to shared
